[#6] YDBSupport hang when gdb used to capture of direct mode process
The hang was fixed by unsetting DEBUGINFOD_URLS before running any program. The fix is inelegant because only `gdb` is affected. But it was the smallest usable fix, and the script basically needs just to work rather than be elegant.

THe issue was found on an [EndeavourOS](https://endeavouros.com/) machine as well as on [Ubuntu 24.04 LTS](https://ubuntu.com/blog/canonical-releases-ubuntu-24-04-noble-numbat) machines. The former has gdb 14.2 and the latter have 15.0.50. Ubuntu 22.04 LTS has 12.1.